MVP - double jump Tutorial
There is now a tutorial for all of the mechanics except the double jump mechanic. The rotate mechanic was changed to having it rotate when a second tap occurs after the player is alrready touching the screen. The Player changes color when the double jump is ready. Fastest time is now at a limit of 4.